+++ src/locales/en_US.ts
@@ -1979,6 +1979,32 @@ const strings = {
 
   // #endregion Home
 
+  // Payment types
+  ach_bank_transfer: 'ACH Bank Transfer',
+  apple_pay: 'Apple Pay',
+  bank_transfer: 'Bank Transfer',
+  in_person_cash_and_debit_card: 'In-person Cash and Debit Card',
+  colombia_bank_transfer: 'Colombia Bank Transfer',
+  credit_and_debit_card: 'Credit and Debit Card',
+  debit_card: 'Debit Card',
+  direct_to_bank: 'Direct to Bank',
+  faster_payments: 'Faster Payments',
+  google_pay: 'Google Pay',
+  instant_ach_bank_transfer: 'Instant ACH Bank Transfer',
+  ideal: 'iDEAL',
+  interac_e_transfer: 'Interac e-Transfer',
+  mexico_bank_transfer: 'Mexico Bank Transfer',
+  payid: 'PayID',
+  paypal: 'Paypal',
+  pix: 'PIX',
+  pse_payment: 'PSE Payment',
+  revolut: 'Revolut',
+  sepa_bank_transfer: 'SEPA Bank Transfer',
+  spei_bank_transfer: 'SPEI Bank Transfer',
+  turkish_bank_transfer: 'Turkish Bank Transfer',
+  venmo: 'Venmo',
+  bank_wire_transfer: 'Bank Wire Transfer',
+
   // #region UK Compliance
 
   uk_deposit_parent_crypto_modal_message_no_exchange_2s: